+++ components/atom/badge/demo/config.js
@@ -1,0 +1,13 @@
+export const extendedColors = [
+  'theblue',
+  'wiggins',
+  'sky',
+  'surfer',
+  'kiwi',
+  'lavender',
+  'sugarcotton',
+  'jungle',
+  'night',
+  'chilli',
+  'apple'
+]