+++ components/atom/panel/demo/config.js
@@ -1,0 +1,13 @@
+export const flexWrapper = {
+  background: 'white',
+  color: 'grey',
+  display: 'flex',
+  flexWrap: 'wrap',
+  justifyContent: 'center'
+}
+
+export const flexItem = {
+  flex: '0 0 auto',
+  margin: '15px',
+  textAlign: 'center'
+}